pyproject.toml
[tool.poetry]
name = "mvp-recruitai"
version = "0.1.0"
description = ""
authors = ["Rithvik-B <rithvik.iitb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
pillow = "^11.1.0"
pdfplumber = "^0.11.5"
openai = "^1.59.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
daphne = "^4.1.2"
celery = "^5.5.2"
redis = "^6.1.0"
boto3 = "^1.38.19"
django-storages = "^1.14.6"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.9.0"
django-celery-results = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
